[Fixed] Crystal tuned attributes to be more appropriate as per live (Changed attackspeed from int to float so might need to delete old crystals) [Added] Crystal / Pearl tuning updates object name
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@4630 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/tangible/component/lightsaber/LightsaberCrystalComponent.h
@@ -120,7 +120,7 @@ public:
 
 	int getSacHealth();
 
-	int getAttackSpeed();
+	float getAttackSpeed();
 
 	int getSacAction();
 
@@ -152,7 +152,7 @@ public:
 
 	void setSacMind(int value);
 
-	void setAttackSpeed(int value);
+	void setAttackSpeed(float value);
 
 	void setWoundChance(float value);
 
@@ -197,7 +197,7 @@ protected:
 
 	String owner;
 
-	int attackSpeed;
+	float attackSpeed;
 
 	int minimumDamage;
 
@@ -250,7 +250,7 @@ public:
 
 	int getSacHealth();
 
-	int getAttackSpeed();
+	float getAttackSpeed();
 
 	int getSacAction();
 
@@ -282,7 +282,7 @@ public:
 
 	void setSacMind(int value);
 
-	void setAttackSpeed(int value);
+	void setAttackSpeed(float value);
 
 	void setWoundChance(float value);
 
@@ -353,7 +353,7 @@ public:
 
 	int getSacHealth();
 
-	int getAttackSpeed();
+	float getAttackSpeed();
 
 	int getSacAction();
 
@@ -385,7 +385,7 @@ public:
 
 	void setSacMind(int value);
 
-	void setAttackSpeed(int value);
+	void setAttackSpeed(float value);
 
 	void setWoundChance(float value);
